requirements.txt
djangorestframework
pyyaml
uritemplate
